Move experimental bundler to default
