--- packages/bundlers/experimental/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@parcel/bundler-experimental",
-  "version": "2.7.0",
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "funding": {
-    "type": "opencollective",
-    "url": "https://opencollective.com/parcel"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/parcel-bundler/parcel.git"
-  },
-  "main": "lib/ExperimentalBundler.js",
-  "source": "src/ExperimentalBundler.js",
-  "engines": {
-    "node": ">= 12.0.0",
-    "parcel": "^2.7.0"
-  },
-  "dependencies": {
-    "@parcel/diagnostic": "2.7.0",
-    "@parcel/graph": "2.7.0",
-    "@parcel/hash": "2.7.0",
-    "@parcel/plugin": "2.7.0",
-    "@parcel/utils": "2.7.0",
-    "nullthrows": "^1.1.1"
-  }
-}
