+++ packages/core/integration-tests/test/bundler.js
@@ -51,53 +51,51 @@ describe('bundler', function () {
   });
 
   it('should remove reused bundle (over shared bundles based on size) if the bundlegroup hit the parallel request limit', async function () {
-    if (process.env.PARCEL_TEST_EXPERIMENTAL_BUNDLER) {
-      let b = await bundle(
-        path.join(
-          __dirname,
-          'integration/shared-bundle-reused-bundle-remove-reuse/index.js',
-        ),
-        {
-          mode: 'production',
-          defaultTargetOptions: {
-            shouldScopeHoist: false,
-          },
+    let b = await bundle(
+      path.join(
+        __dirname,
+        'integration/shared-bundle-reused-bundle-remove-reuse/index.js',
+      ),
+      {
+        mode: 'production',
+        defaultTargetOptions: {
+          shouldScopeHoist: false,
         },
-      );
+      },
+    );
 
-      assertBundles(b, [
-        {
-          name: 'index.js',
-          assets: [
-            'index.js',
-            'bundle-url.js',
-            'cacheLoader.js',
-            'css-loader.js',
-            'esmodule-helpers.js',
-            'js-loader.js',
-            'bundle-manifest.js',
-          ],
-        },
-        {
-          assets: ['bar.js', 'foo.js', 'a.js', 'b.js'],
-        },
-        {
-          assets: ['buzz.js'],
-        },
-        {
-          assets: ['c.js'],
-        },
-        {
-          assets: ['a.js', 'b.js', 'foo.js'],
-        },
-        {
-          assets: ['styles.css'],
-        },
-        {
-          assets: ['local.html'],
-        },
-      ]);
-    }
+    assertBundles(b, [
+      {
+        name: 'index.js',
+        assets: [
+          'index.js',
+          'bundle-url.js',
+          'cacheLoader.js',
+          'css-loader.js',
+          'esmodule-helpers.js',
+          'js-loader.js',
+          'bundle-manifest.js',
+        ],
+      },
+      {
+        assets: ['bar.js', 'foo.js', 'a.js', 'b.js'],
+      },
+      {
+        assets: ['buzz.js'],
+      },
+      {
+        assets: ['c.js'],
+      },
+      {
+        assets: ['a.js', 'b.js', 'foo.js'],
+      },
+      {
+        assets: ['styles.css'],
+      },
+      {
+        assets: ['local.html'],
+      },
+    ]);
   });
 
   //This test case is the same as previous except we remove the shared bundle since it is smaller
@@ -205,60 +203,56 @@ describe('bundler', function () {
     //The shared bundle should only be 'put back' for the bundlegroups which hit the parallel request limit
     // But if there are at least two other bundlegroups using this shared bundle that do not hit the max limit
     // the shared bundle should not be removed from the graph
-    if (process.env.PARCEL_TEST_EXPERIMENTAL_BUNDLER) {
-      let b = await bundle(
-        path.join(
-          __dirname,
-          'integration/shared-bundle-between-reused-bundle-removal/index.js',
-        ),
-        {
-          mode: 'production',
-          defaultTargetOptions: {
-            shouldScopeHoist: false,
-          },
+    let b = await bundle(
+      path.join(
+        __dirname,
+        'integration/shared-bundle-between-reused-bundle-removal/index.js',
+      ),
+      {
+        mode: 'production',
+        defaultTargetOptions: {
+          shouldScopeHoist: false,
         },
-      );
+      },
+    );
 
-      assertBundles(b, [
-        {
-          name: 'index.js',
-          assets: [
-            'index.js',
-            'bundle-url.js',
-            'cacheLoader.js',
-            'css-loader.js',
-            'esmodule-helpers.js',
-            'js-loader.js',
-            'bundle-manifest.js',
-          ],
-        },
-        {
-          assets: ['bar.js', 'foo.js', 'a.js', 'b.js'], // shared bundle merged back
-        },
-        {
-          assets: ['buzz.js'],
-        },
-        {
-          assets: ['c.js'], // shared bundle
-        },
-        {
-          assets: ['foo.js', 'a.js', 'b.js'],
-        },
-        {
-          assets: ['styles.css'],
-        },
-        {
-          assets: ['local.html'],
-        },
-      ]);
+    assertBundles(b, [
+      {
+        name: 'index.js',
+        assets: [
+          'index.js',
+          'bundle-url.js',
+          'cacheLoader.js',
+          'css-loader.js',
+          'esmodule-helpers.js',
+          'js-loader.js',
+          'bundle-manifest.js',
+        ],
+      },
+      {
+        assets: ['bar.js', 'foo.js', 'a.js', 'b.js'], // shared bundle merged back
+      },
+      {
+        assets: ['buzz.js'],
+      },
+      {
+        assets: ['c.js'], // shared bundle
+      },
+      {
+        assets: ['foo.js', 'a.js', 'b.js'],
+      },
+      {
+        assets: ['styles.css'],
+      },
+      {
+        assets: ['local.html'],
+      },
+    ]);
 
-      assert(
-        b
-          .getReferencedBundles(
-            b.getBundlesWithAsset(findAsset(b, 'bar.js'))[0],
-          )
-          .includes(b.getBundlesWithAsset(findAsset(b, 'c.js'))[0]),
-      );
-    }
+    assert(
+      b
+        .getReferencedBundles(b.getBundlesWithAsset(findAsset(b, 'bar.js'))[0])
+        .includes(b.getBundlesWithAsset(findAsset(b, 'c.js'))[0]),
+    );
   });
 });
